updated in myems-api/core/point.py

+++ myems-api/core/point.py
@@ -583,6 +583,68 @@ class PointItem:
                                    title='API.BAD_REQUEST',
                                    description='API.THERE_IS_RELATION_WITH_VIRTUAL_POWER_PLANTS')
 
+        # check if this point is being used by energy storage container battery
+        cursor.execute(" SELECT name "
+                       " FROM tbl_energy_storage_containers_batteries "
+                       " WHERE battery_state_point_id = %s "
+                       "   OR soc_point_id = %s "
+                       "   OR power_point_id = %s "
+                       " LIMIT 1 ",
+                        (id_, id_, id_,))
+        row_energy_storage_container_battery = cursor.fetchone()
+        if row_energy_storage_container_battery is not None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_400,
+                                    title='API.BAD_REQUEST',
+                                    description='API.THERE_IS_RELATION_WITH_ENERGY_STORAGE_CONTAINERS_BATTERIES')
+
+        # check if this point is being used by energy storage container grid
+        cursor.execute(" SELECT name "
+                       " FROM tbl_energy_storage_containers_grids "
+                       " WHERE power_point_id = %s "
+                       " LIMIT 1 ",
+                       (id_,))
+        row_energy_storage_container_grid = cursor.fetchone()
+        if row_energy_storage_container_grid is not None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_400,
+                                   title='API.BAD_REQUEST',
+                                   description='API.THERE_IS_RELATION_WITH_ENERGY_STORAGE_CONTAINERS_GRIDS')
+
+        # check if this point is being used by energy storage container load
+        cursor.execute(" SELECT name "
+                       " FROM tbl_energy_storage_containers_loads "
+                       " WHERE power_point_id = %s "
+                       " LIMIT 1 ",
+                       (id_,))
+        row_energy_storage_container_load = cursor.fetchone()
+        if row_energy_storage_container_load is not None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_400,
+                                   title='API.BAD_REQUEST',
+                                   description='API.THERE_IS_RELATION_WITH_ENERGY_STORAGE_CONTAINERS_LOADS')
+
+        # check if this point is being used by energy storage container power conversion system
+        cursor.execute(" SELECT name "
+                       " FROM tbl_energy_storage_containers_power_conversion_systems "
+                       " WHERE run_state_point_id = %s "
+                       "   OR today_charge_energy_point_id = %s "
+                       "   OR today_discharge_energy_point_id = %s "
+                       "   OR total_charge_energy_point_id = %s "
+                       "   OR total_discharge_energy_point_id = %s "
+                       " LIMIT 1 ",
+                       (id_, id_, id_, id_, id_,))
+        row_energy_storage_container_power_conversion_system = cursor.fetchone()
+        if row_energy_storage_container_power_conversion_system is not None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_400,
+                                   title='API.BAD_REQUEST',
+                                   description='API.THERE_IS_RELATION_WITH_ENERGY_STORAGE_CONTAINERS_POWER_CONVERSION_SYSTEMS')
+
         cursor.execute(" DELETE FROM tbl_points WHERE id = %s ", (id_,))
         cnx.commit()
